[HU-INTEGRACION] Michael Menendez: Reparar integracion de ramas - Intercambia contenido de Models/Anuncio.cs y Controllers/AnunciosController.cs, restaura DbSets Cuenta/Anuncio en ApplicationDbContext, unifica Usuario como perfil ligado a Cuenta (IdUsuario/IdCuenta/Nombre) y ajusta FKs IdUsuario a long, regenera la cadena de migraciones como una sola migracion Inicial con datos semilla consistentes.

+++ HuellitasSV.API/Models/SolicitudAdopcion.cs
@@ -35,7 +35,7 @@ public class SolicitudAdopcion
     /// <summary>
     /// Identificador del usuario solicitante (adoptante).
     /// </summary>
-    public int IdUsuario { get; set; }
+    public long IdUsuario { get; set; }
 
     /// <summary>
     /// Nombre de contacto confirmado por el usuario al enviar la solicitud.
